Jeehea : Handle HTTP status code

+++ src/main/java/com/weave/weaveserver/service/ArchiveService.java
@@ -1,10 +1,14 @@
 package com.weave.weaveserver.service;
 
+import com.weave.weaveserver.config.exception.ConflictException;
+import com.weave.weaveserver.config.exception.ForbiddenException;
+import com.weave.weaveserver.config.exception.NotFoundException;
 import com.weave.weaveserver.config.jwt.TokenService;
 import com.weave.weaveserver.domain.*;
 import com.weave.weaveserver.dto.*;
 import com.weave.weaveserver.repository.*;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.annotations.NotFound;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -26,19 +30,38 @@ public class ArchiveService {
 
     public final UserRepository userRepository;
     public final TeamRepository teamRepository;
+    public final BelongRepository belongRepository;
     public final CategoryRepository categoryRepository;
     public final ArchiveRepository archiveRepository;
     public final ImageRepository imageRepository;
 
 
     public void addArchive(ArchiveRequest.createRequest request, HttpServletRequest servletRequest){
-        String userEmail = tokenService.getUserEmail(servletRequest); // 토큰으로부터 user 이메일 가져오기
-        User user = userRepository.findUserByEmail(userEmail);
+        User clientUser = findUserByEmailInToken(servletRequest);
 
-        Team team = teamRepository.getReferenceById(request.getTeamIdx());
-        Category category = categoryRepository.getReferenceById(request.getCategoryIdx());
+        //Team team = teamRepository.getReferenceById(request.getTeamIdx()); //TODO : team없을 때 이 에러 잡는 법 모르겠음!! SQL단위 에러인 듯
+        Team team = teamRepository.findByTeamIdx(request.getTeamIdx());
+        if(team == null){
+            System.out.println("jh : team == null");
+            throw new ConflictException("Team is not found by the teamIdx in request body");
+        }
+
+        checkBelong(team.getTeamIdx(), clientUser.getEmail());
+
+//        Category category = categoryRepository.getReferenceById(request.getCategoryIdx()); //TODO : 동일
+        Category category = categoryRepository.findByCategoryIdx(request.getCategoryIdx());
+        if(category == null){
+            System.out.println("jh : category == null");
+            throw new ConflictException("Category is not found by the categoryIdx in request body");
+        }
+
+        if(request.getTitle().isBlank()){ //""인지 + 공백으로만 된 문자열인지 검사
+            System.out.println("jh : title is empty or blank");
+            throw new ConflictException("Title of archive cannot be empty or blank");
+        }
+
         Archive archive = Archive.builder()
-                .user(user)
+                .user(clientUser)
                 .team(team)
                 .title(request.getTitle())
                 .content(request.getContent())
@@ -51,9 +74,18 @@ public class ArchiveService {
     }
 
     @Transactional // 왜 이걸 붙이면 LAZY 관련 에러가 해결되는 거지?
-    public ArchiveResponse.archiveListResponseContainer getArchiveList(Long teamIdx){
+    public ArchiveResponse.archiveListResponseContainer getArchiveList(Long teamIdx, HttpServletRequest servletRequest){
+        User clientUser = findUserByEmailInToken(servletRequest);
+
         //Team
         Team team = teamRepository.findByTeamIdx(teamIdx);
+        if(team == null){
+            System.out.println("jh : team == null");
+            throw new NotFoundException("Team is not found by this teamIdx");
+        }
+
+        checkBelong(team.getTeamIdx(), clientUser.getEmail());
+
         List<LocalDate> dateList = team.getStartDate().datesUntil(team.getEndDate().plusDays(1))
                 .collect(Collectors.toList());
         TeamResponse.teamWithDateListResponse teamResponse = new TeamResponse.teamWithDateListResponse(
@@ -76,10 +108,10 @@ public class ArchiveService {
             Long archiveIdx = a.getArchiveIdx();
 
             //User
-            User user = a.getUser();
+            User archiveUser = a.getUser();
             UserResponse.userResponse userResponse = new UserResponse.userResponse(
-                    user.getName(),
-                    user.getEmail()
+                    archiveUser.getName(),
+                    archiveUser.getEmail()
             );
             userList.put(archiveIdx, userResponse);
 
@@ -118,14 +150,23 @@ public class ArchiveService {
 
 
     @Transactional // 왜 이걸 붙이면 LAZY 관련 에러가 해결되는 거지?
-    public ArchiveResponse.archiveResponse getArchiveDetail(Long archiveIdx){
+    public ArchiveResponse.archiveResponse getArchiveDetail(Long archiveIdx, HttpServletRequest servletRequest){
+        User clientUser = findUserByEmailInToken(servletRequest);
+
         Archive archive = archiveRepository.findByArchiveIdx(archiveIdx);
+        if(archive == null){
+            System.out.println("jh : archive == null");
+            throw new NotFoundException("Archive is not found by this archiveIdx");
+        }
+
+        Team team = archive.getTeam();
+        checkBelong(team.getTeamIdx(), clientUser.getEmail());
 
         //User
-        User user = archive.getUser();
+        User archiveUser = archive.getUser();
         UserResponse.userResponse userResponse = new UserResponse.userResponse(
-                user.getName(),
-                user.getEmail()
+                archiveUser.getName(),
+                archiveUser.getEmail()
         );
 
         //ImageList
@@ -154,17 +195,54 @@ public class ArchiveService {
         return response;
     }
 
-    public void updateArchivePin(Long archiveIdx){
+    public void updateArchivePin(Long archiveIdx, HttpServletRequest servletRequest){
+        User clientUser = findUserByEmailInToken(servletRequest);
+
         Archive archive = archiveRepository.findByArchiveIdx(archiveIdx);
+        if(archive == null){
+            System.out.println("jh : archive == null");
+            throw new NotFoundException("Archive is not found by this archiveIdx");
+        }
+
+        Team team = archive.getTeam();
+        checkBelong(team.getTeamIdx(), clientUser.getEmail());
+
         archive.updateArchive(false);
         archiveRepository.save(archive);
     }
 
     @Transactional
-    public void deleteArchive(Long archiveIdx){
+    public void deleteArchive(Long archiveIdx, HttpServletRequest servletRequest){
+        User clientUser = findUserByEmailInToken(servletRequest);
+
         Archive archive = archiveRepository.findByArchiveIdx(archiveIdx);
+        if(archive != null) { //archiveIdx에 해당하는 archive가 존재할 때만 belong체크와 삭제 실행
+            Team team = archive.getTeam();
+            checkBelong(team.getTeamIdx(), clientUser.getEmail());
+
 //        imageRepository.deleteByArchive(archive);
-        archiveRepository.deleteByArchiveIdx(archiveIdx);
+            archiveRepository.deleteByArchiveIdx(archiveIdx);
+        }
+        else {
+            System.out.println("jh : archive == null. No delete");
+        }
+
+    }
+
+//    private void throwNotFoundException(String )
+    private User findUserByEmailInToken(HttpServletRequest servletRequest){
+        String userEmail = tokenService.getUserEmail(servletRequest); // 토큰으로부터 user 이메일 가져오기
+        User clientUser = userRepository.findUserByEmail(userEmail);
+
+        return clientUser;
+    }
+
+    private void checkBelong(Long teamIdx, String email){
+        Belong belong = belongRepository.findByTeamIdxAndUser(teamIdx, email);
+        if(belong == null){
+            System.out.println("jh : belong == null");
+            throw new ForbiddenException("Forbidden. User is not belong in the team");
+        }
     }
 
 }
